tests/wpt/tests/html/syntax/parsing/resources/foreign-fragment.dat

#data
<nobr>X
#errors
6: HTML start tag “nobr” in a foreign namespace context.
7: End of file seen and there were open elements.
#document-fragment
svg path
#document
| <nobr>
|   "X"

#data
<font color></font>X
#errors
12: HTML start tag “font” in a foreign namespace context.
#document-fragment
svg path
#document
| <font>
|   color=""
| "X"

#data
<font></font>X
#errors
#document-fragment
svg path
#document
| <svg font>
| "X"

#data
<g></path>X
#errors
10: End tag “path” did not match the name of the current open element (“g”).
11: End of file seen and there were open elements.
#document-fragment
svg path
#document
| <svg g>
|   "X"

#data
</path>X
#errors
5: Stray end tag “path”.
#document-fragment
svg path
#document
| "X"

#data
</foreignObject>X
#errors
5: Stray end tag “foreignobject”.
#document-fragment
svg foreignObject
#document
| "X"

#data
</desc>X
#errors
5: Stray end tag “desc”.
#document-fragment
svg desc
#document
| "X"

#data
</title>X
#errors
5: Stray end tag “title”.
#document-fragment
svg title
#document
| "X"

#data
</svg>X
#errors
5: Stray end tag “svg”.
#document-fragment
svg svg
#document
| "X"

#data
</mfenced>X
#errors
5: Stray end tag “mfenced”.
#document-fragment
math mfenced
#document
| "X"

#data
</malignmark>X
#errors
5: Stray end tag “malignmark”.
#document-fragment
math malignmark
#document
| "X"

#data
</math>X
#errors
5: Stray end tag “math”.
#document-fragment
math math
#document
| "X"

#data
</annotation-xml>X
#errors
5: Stray end tag “annotation-xml”.
#document-fragment
math annotation-xml
#document
| "X"

#data
</mtext>X
#errors
5: Stray end tag “mtext”.
#document-fragment
math mtext
#document
| "X"

#data
</mi>X
#errors
5: Stray end tag “mi”.
#document-fragment
math mi
#document
| "X"

#data
</mo>X
#errors
5: Stray end tag “mo”.
#document-fragment
math mo
#document
| "X"

#data
</mn>X
#errors
5: Stray end tag “mn”.
#document-fragment
math mn
#document
| "X"

#data
</ms>X
#errors
5: Stray end tag “ms”.
#document-fragment
math ms
#document
| "X"

#data
<b></b><mglyph/><i></i><malignmark/><u></u><ms/>X
#errors
51: Self-closing syntax (“/>”) used on a non-void HTML element. Ignoring the slash and treating as a start tag.
52: End of file seen and there were open elements.
#new-errors
(1:44-1:49) non-void-html-element-start-tag-with-trailing-solidus
#document-fragment
math ms
#document
| <b>
| <math mglyph>
| <i>
| <math malignmark>
| <u>
| <ms>
|   "X"

#data
<malignmark></malignmark>
#errors
#document-fragment
math ms
#document
| <math malignmark>

#data
<div></div>
#errors
#document-fragment
math ms
#document
| <div>

#data
<figure></figure>
#errors
#document-fragment
math ms
#document
| <figure>

#data
<b></b><mglyph/><i></i><malignmark/><u></u><mn/>X
#errors
51: Self-closing syntax (“/>”) used on a non-void HTML element. Ignoring the slash and treating as a start tag.
52: End of file seen and there were open elements.
#new-errors
(1:44-1:49) non-void-html-element-start-tag-with-trailing-solidus
#document-fragment
math mn
#document
| <b>
| <math mglyph>
| <i>
| <math malignmark>
| <u>
| <mn>
|   "X"

#data
<malignmark></malignmark>
#errors
#document-fragment
math mn
#document
| <math malignmark>

#data
<div></div>
#errors
#document-fragment
math mn
#document
| <div>

#data
<figure></figure>
#errors
#document-fragment
math mn
#document
| <figure>

#data
<b></b><mglyph/><i></i><malignmark/><u></u><mo/>X
#errors
51: Self-closing syntax (“/>”) used on a non-void HTML element. Ignoring the slash and treating as a start tag.
52: End of file seen and there were open elements.
#new-errors
(1:44-1:49) non-void-html-element-start-tag-with-trailing-solidus
#document-fragment
math mo
#document
| <b>
| <math mglyph>
| <i>
| <math malignmark>
| <u>
| <mo>
|   "X"

#data
<malignmark></malignmark>
#errors
#document-fragment
math mo
#document
| <math malignmark>

#data
<div></div>
#errors
#document-fragment
math mo
#document
| <div>

#data
<figure></figure>
#errors
#document-fragment
math mo
#document
| <figure>

#data
<b></b><mglyph/><i></i><malignmark/><u></u><mi/>X
#errors
51: Self-closing syntax (“/>”) used on a non-void HTML element. Ignoring the slash and treating as a start tag.
52: End of file seen and there were open elements.
#new-errors
(1:44-1:49) non-void-html-element-start-tag-with-trailing-solidus
#document-fragment
math mi
#document
| <b>
| <math mglyph>
| <i>
| <math malignmark>
| <u>
| <mi>
|   "X"

#data
<malignmark></malignmark>
#errors
#document-fragment
math mi
#document
| <math malignmark>

#data
<div></div>
#errors
#document-fragment
math mi
#document
| <div>

#data
<figure></figure>
#errors
#document-fragment
math mi
#document
| <figure>

#data
<b></b><mglyph/><i></i><malignmark/><u></u><mtext/>X
#errors
51: Self-closing syntax (“/>”) used on a non-void HTML element. Ignoring the slash and treating as a start tag.
52: End of file seen and there were open elements.
#new-errors
(1:44-1:52) non-void-html-element-start-tag-with-trailing-solidus
#document-fragment
math mtext
#document
| <b>
| <math mglyph>
| <i>
| <math malignmark>
| <u>
| <mtext>
|   "X"

#data
<malignmark></malignmark>
#errors
#document-fragment
math mtext
#document
| <math malignmark>

#data
<div></div>
#errors
#document-fragment
math mtext
#document
| <div>

#data
<figure></figure>
#errors
#document-fragment
math mtext
#document
| <figure>

#data
<div></div>
#errors
5: HTML start tag “div” in a foreign namespace context.
#document-fragment
math annotation-xml
#document
| <div>

#data
<figure></figure>
#errors
#document-fragment
math annotation-xml
#document
| <math figure>

#data
<div></div>
#errors
5: HTML start tag “div” in a foreign namespace context.
#document-fragment
math math
#document
| <div>

#data
<figure></figure>
#errors
#document-fragment
math math
#document
| <math figure>

#data
<div></div>
#errors
#document-fragment
svg foreignObject
#document
| <div>

#data
<figure></figure>
#errors
#document-fragment
svg foreignObject
#document
| <figure>

#data
<div></div>
#errors
#document-fragment
svg title
#document
| <div>

#data
<figure></figure>
#errors
#document-fragment
svg title
#document
| <figure>

#data
<figure></figure>
#errors
#document-fragment
svg desc
#document
| <figure>

#data
<div><h1>X</h1></div>
#errors
5: HTML start tag “div” in a foreign namespace context.
#document-fragment
svg svg
#document
| <div>
|   <h1>
|     "X"

#data
<div></div>
#errors
5: HTML start tag “div” in a foreign namespace context.
#document-fragment
svg svg
#document
| <div>

#data
<div></div>
#errors
#document-fragment
svg desc
#document
| <div>

#data
<plaintext><foo>
#errors
(1,16): expected-closing-tag-but-got-eof
#document-fragment
svg desc
#document
| <plaintext>
|   "<foo>"

#data
<frameset>X
#errors
6: Stray start tag “frameset”.
#document-fragment
svg desc
#document
| "X"

#data
<head>X
#errors
6: Stray start tag “head”.
#document-fragment
svg desc
#document
| "X"

#data
<body>X
#errors
6: Stray start tag “body”.
#document-fragment
svg desc
#document
| "X"

#data
<html>X
#errors
6: Stray start tag “html”.
#document-fragment
svg desc
#document
| "X"

#data
<html class="foo">X
#errors
6: Stray start tag “html”.
#document-fragment
svg desc
#document
| "X"

#data
<body class="foo">X
#errors
6: Stray start tag “body”.
#document-fragment
svg desc
#document
| "X"

#data
<svg><p>
#errors
8: HTML start tag “p” in a foreign namespace context.
#document-fragment
div
#document
| <svg svg>
| <p>

#data
<p>
#errors
3: HTML start tag “p” in a foreign namespace context.
#document-fragment
svg svg
#document
| <p>

#data
<svg></p><foo>
#errors
9: HTML end tag “p” in a foreign namespace context.
(1:6) Unexpected </p> from in body insertion mode
(1:15) Unexpected EOF
#document-fragment
div
#document
| <svg svg>
| <p>
| <foo>

#data
<svg></br><foo>
#errors
10: HTML end tag “br” in a foreign namespace context.
(1:6) Unexpected </br> from in body insertion mode
(1:16) Unexpected EOF
#document-fragment
div
#document
| <svg svg>
| <br>
| <foo>

#data
</p><foo>
#errors
4: HTML end tag “p” in a foreign namespace context.
(1:1) Unexpected </p> from in body insertion mode
(1:10) Unexpected EOF
#document-fragment
svg svg
#document
| <p>
| <svg foo>

#data
</br><foo>
#errors
5: HTML end tag “br” in a foreign namespace context.
(1:1) Unexpected </br> from in body insertion mode
(1:11) Unexpected EOF
#document-fragment
svg svg
#document
| <br>
| <svg foo>

#data
<body><foo>
#errors
6: HTML start tag “body” in a foreign namespace context.
(1:1) Unexpected <body> from in body insertion mode
(1:12) Unexpected EOF
#document-fragment
svg svg
#document
| <svg foo>

#data
<p><foo>
#errors
3: HTML start tag “p” in a foreign namespace context.
(1:9) Unexpected EOF
#document-fragment
svg svg
#document
| <p>
|   <foo>

#data
<p></p><foo>
#errors
3: HTML start tag “p” in a foreign namespace context.
(1:13) Unexpected EOF
#document-fragment
svg svg
#document
| <p>
| <svg foo>
